requirements.txt
flask==3.1.2
rapidfuzz==3.14.1
webdavclient3==3.14.6
python-dotenv==1.1.1
psycopg2-binary==2.9.10
cryptography==46.0.1
prettytable==3.16.0
customtkinter==5.2.2
flask-cors==6.0.1
datetime==5.5
flask-limiter==3.13